refactor: centralize Utila keychain config
