+++ apps/sdp-api/src/services/domain/signing/provider-adapter-factory.ts
@@ -11,6 +11,7 @@ import {
   KeychainUtilaAdapter,
   type SigningConfigRecord,
 } from "@/services/adapters";
+import { buildKeychainUtilaConfig } from "@/services/adapters/signing/keychain/utila-config";
 import { createDfnsApiClient, normalizeDfnsWalletId } from "@/services/dfns/client";
 import { createEncryptionService } from "@/services/encryption.service";
 import {
@@ -247,34 +248,14 @@ const providerAdapterFactories = {
   },
   anchorage: async () => new LifecycleOnlyAdapter("anchorage"),
   utila: async ({ env, record, parsed }) => {
-    const serviceAccountEmail = env.UTILA_SERVICE_ACCOUNT_EMAIL;
-    const serviceAccountPrivateKeyPem = env.UTILA_SERVICE_ACCOUNT_PRIVATE_KEY;
-    const vaultId = parsed.vaultId ?? env.UTILA_VAULT_ID;
-
-    if (!serviceAccountEmail || !serviceAccountPrivateKeyPem || !vaultId) {
-      throw new SigningError(
-        "Utila configuration is missing service account credentials or vault ID",
-        "PROVIDER_NOT_CONFIGURED"
-      );
-    }
-
-    return new KeychainUtilaAdapter({
-      serviceAccountEmail,
-      serviceAccountPrivateKeyPem,
-      vaultId,
-      network: resolveUtilaNetwork(env, parsed.network),
-      apiBaseUrl: parsed.apiBaseUrl ?? env.UTILA_API_BASE_URL,
-      pollIntervalMs: parseOptionalPositiveInteger(
-        env.UTILA_POLL_INTERVAL_MS,
-        "UTILA_POLL_INTERVAL_MS"
-      ),
-      maxPollAttempts: parseOptionalPositiveInteger(
-        env.UTILA_MAX_POLL_ATTEMPTS,
-        "UTILA_MAX_POLL_ATTEMPTS"
-      ),
-      designatedSigners: parseCsvList(env.UTILA_DESIGNATED_SIGNERS),
-      defaultWalletId: record.defaultWalletId ?? undefined,
-    });
+    return new KeychainUtilaAdapter(
+      buildKeychainUtilaConfig(env, {
+        apiBaseUrl: parsed.apiBaseUrl,
+        defaultWalletId: record.defaultWalletId,
+        network: parsed.network,
+        vaultId: parsed.vaultId,
+      })
+    );
   },
 } satisfies {
   [K in ProviderConfigRecord["provider"]]: AdapterFactory<
@@ -290,39 +271,4 @@ export async function createAdapterFromEncryptedConfig(
   const parsed = await parseConfigRecord(env, orgId, record);
   const factory = providerAdapterFactories[parsed.provider] as AdapterFactory;
   return factory({ env, orgId, record, parsed });
-}
-
-function resolveUtilaNetwork(
-  env: Env,
-  configured?: "networks/solana-mainnet" | "networks/solana-devnet"
-): "networks/solana-mainnet" | "networks/solana-devnet" {
-  if (configured) {
-    return configured;
-  }
-  if (env.UTILA_NETWORK) {
-    return env.UTILA_NETWORK;
-  }
-  return env.SOLANA_NETWORK === "mainnet-beta"
-    ? "networks/solana-mainnet"
-    : "networks/solana-devnet";
-}
-
-function parseOptionalPositiveInteger(
-  value: string | undefined,
-  envVarName: string
-): number | undefined {
-  if (!value) return undefined;
-  const parsed = Number(value);
-  if (!Number.isInteger(parsed) || parsed <= 0) {
-    throw new SigningError(`${envVarName} must be a positive integer`, "INVALID_REQUEST");
-  }
-  return parsed;
-}
-
-function parseCsvList(value: string | undefined): string[] | undefined {
-  const items = value
-    ?.split(",")
-    .map((item) => item.trim())
-    .filter(Boolean);
-  return items && items.length > 0 ? items : undefined;
 }
